requirements.txt
pyyaml>=5.1
sentencepiece
